requirements.txt
Django==1.7.4
django-autofixture==0.9.1
Markdown==2.5.2
astroid==1.3.4
django-pagedown==0.1.0
jedi==0.8.1-final0
logilab-common==0.63.2
six==1.9.0
pytz==2014.10
